UefiPayloadPkg: Fix bug that Elf loader doesn't relocate in some case
Current implementation of the Elf loader reuses the same memory range if
the desired memory range is covered by [file base, file base + file size].
However, there is a potentil bug, for example:
If the Elf file is located at 0x1000,(which means Elf header is at 0x1000),
and the the first segment is located at 0x2000.
When we parse and try to load the elf file, if the first segment's desire
address is 0x1000, current implementation may load it into 0x1000, and
override the Elf header. Similiarly, it may also override other segment.

Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>